add map_helpers
